Move 'BinaryMessage' to dedicated file
